Disable anki connect when connections fail
This is a minor hack due to the multi-second delay calling a closed
localhost port on Windows. There are better ways to solve this (not
least just rendering the card stuff separately), but this is simple
for now.

This commit also includes very minor text changes to headings, since
the side panel looks very cramped by default with the old "definitions"
and "sentences" text.